// src/app/type.d.ts
export type MaterialConfigId = string;
export type FinishGroupId = string;
export type MaterialGroupId = string;
export type MaterialId = string;
export type QuoteId = string;
export type VendorId = string;
export type ConfigId = string;
export type ShippingId = string;
export type ModelId = string;
export type ModelSceneId = string;
export type FileId = string;
export type PriceId = string;
export type CartId = string;
export type ConfigurationId = string;
export type OrderId = string;
export type PaymentId = string;
export type UserId = string;

// Material structure json-schema
// https://github.com/all3dp/material-structure/blob/master/src/schema.js

export type Notification = {
  message: string;
  warning?: boolean;
};

export type PrintingService = {
  [vendorId: string]: {
    materialId: MaterialId;
    finishId: FinishGroupId;
    printingMethodShort: string;
    printingMethod: string;
    materialName: string;
    productionTimeFast: number;
    productionTimeSlow: number;
  };
};

export type MaterialConfig = {
  id: MaterialConfigId;
  name: string;
  color: string;
  colorCode: string;
  colorImage: string;
  printingService: PrintingService;
  finishGroupId: FinishGroupId;
  materialId: MaterialId;
  materialGroupId: MaterialGroupId;
};
// ...
export type UploadingFile = {
  fileId: FileId;
  fileName: string;
  fileSize: number;
  progress: number;
  error: boolean;
  errorMessage?: string;
};

export type BackendModel = {
  modelId: ModelId;
  fileName: string;
  fileUnit: "mm" | "cm" | "in";
  area?: number;
  volume?: number;
  dimensions?: {
    x: number;
    y: number;
    z: number;
  };
  thumbnailUrl: string;
  sceneId?: ModelSceneId;
};
// ...
export type ModelConfigUploading = {
  type: "UPLOADING";
  fileId: FileId;
  id: ConfigId;
};

export type ModelConfigUploaded = {
  type: "UPLOADED";
  quantity: number;
  modelId: ModelId;
  id: ConfigId;
  quoteId?: QuoteId;
  shippingId?: ShippingId;
};

export type ModelConfig = ModelConfigUploading | ModelConfigUploaded;